Stage 2 usage-ingest — three CI failures from first postgres run
Full-read audit surfaced four failures. Three fixed here; the fourth
(111 partition-count expected 2, CI shows 12) needs local Postgres
debug and stays for Cursor.

1. composite-fk.test.js:76 — CORRECTION with mismatched composite
   pair fails with 23503 (FK), not 23514. Cursor's own DL-061 says
   as much. Assertion corrected to 23503. The second sub-case
   already expects 23503 correctly.

2. dlq-worker.js:100-104 — UPDATE used $2 both as `attempts = $2`
   (INTEGER context) and inside `power(2, $2)` where PG can't infer
   the type — "inconsistent types deduced for parameter $2". Cast
   both uses to $2::integer.

3. migrations/111_fin_usage_events.sql apply_usage_events_security —
   short-circuited before ENABLE + FORCE ROW LEVEL SECURITY for
   partition tables. `usage_events_default` therefore had RLS
   DISABLED, tripping the pre-existing H12 force-rls test.
   Moved ENABLE + FORCE above the parent-only branch so every
   partition also gets FORCE RLS (policies still live only on the
   parent — PostgreSQL rejects CREATE POLICY on partitions, so a
   partition with FORCE RLS and no policies denies direct
   non-owner access, which is the desired defense in depth against
   direct-to-partition tenant leaks).

Remaining CI failure — 111 partition test expects 2 partitions,
CI reports 12. Root cause is not visible from static reading:
seedWorld inserts one legal entity (WC-KSA → usage_events_ksa)
plus usage_events_default. The migration's DO $$ loop runs at
migration time when platform_legal_entities is empty. Where 10
extra partitions come from is a runtime mystery — needs Cursor
local Postgres debug (probably a test-isolation or template-DB
artifact).

live P0s: not silently patched. commercial.* / postgres-adapter.js /
current audit_log writers unchanged.

+++ backend/src/fin/usage/dlq-worker.js
@@ -98,10 +98,10 @@ export async function runUsageDlqWorker(pool, {
         const dead = nextAttempts >= MAX_ATTEMPTS
         await lockClient.query(
           `UPDATE fin.usage_events_dlq
-              SET attempts = $2,
+              SET attempts = $2::integer,
                   last_attempt_at = $3::timestamptz,
                   next_retry_at = $3::timestamptz
-                    + (interval '1 second' * (60 * power(2, $2))),
+                    + (interval '1 second' * (60 * power(2, $2::integer))),
                   error_code = COALESCE($4, error_code),
                   dead_lettered_at = CASE WHEN $5 THEN $3::timestamptz ELSE dead_lettered_at END,
                   updated_at = $3::timestamptz
